Share calc_sample_n between mode and delta sampling
