+++ pco/src/data_types/float.rs
@@ -89,7 +89,7 @@ fn choose_mode_bids<F: Float>(nums: &[F], chunk_config: &ChunkConfig) -> PcoResu
         split_fn: Box::new(|nums| classic::split_latents(nums)),
       }];
 
-      if let Some(sample) = sampling::choose_sample(nums, filter_sample) {
+      if let Some(sample) = sampling::choose_mode_sample(nums, filter_sample) {
         bids.extend(float_mult::compute_bid(&sample));
         bids.extend(float_quant::compute_bid(&sample));
       }
